UI: Remove old promql editor (#4864)
